Fix package name from core to com
COSDK-488
