+++ core/src/test/java/com/adyen/checkout/core/internal/ui/state/TestDelegateState.kt
@@ -1,0 +1,13 @@
+/*
+ * Copyright (c) 2025 Adyen N.V.
+ *
+ * This file is open source and available under the MIT license. See the LICENSE file for more info.
+ *
+ * Created by ararat on 5/3/2025.
+ */
+
+package com.adyen.checkout.core.internal.ui.state
+
+internal class TestDelegateState(
+    override val isValid: Boolean = true
+) : DelegateState
