--- core/src/test/java/core/adyen/checkout/core/internal/ui/state/TestDelegateStateFactory.kt
@@ -1,18 +1,0 @@
-/*
- * Copyright (c) 2025 Adyen N.V.
- *
- * This file is open source and available under the MIT license. See the LICENSE file for more info.
- *
- * Created by ararat on 6/3/2025.
- */
-
-package core.adyen.checkout.core.internal.ui.state
-
-import com.adyen.checkout.core.internal.ui.state.DelegateStateFactory
-
-internal class TestDelegateStateFactory(
-    private val defaultDelegateState: TestDelegateState
-) : DelegateStateFactory<TestDelegateState, TestFieldId> {
-    override fun createDefaultDelegateState(): TestDelegateState = defaultDelegateState
-    override fun getFieldIds(): List<TestFieldId> = TestFieldId.entries.toList()
-}
