--- core/src/test/java/core/adyen/checkout/core/internal/ui/state/TestFieldId.kt
@@ -1,17 +1,0 @@
-/*
- * Copyright (c) 2025 Adyen N.V.
- *
- * This file is open source and available under the MIT license. See the LICENSE file for more info.
- *
- * Created by ararat on 6/3/2025.
- */
-
-package core.adyen.checkout.core.internal.ui.state
-
-import com.adyen.checkout.core.internal.ui.state.model.FieldId
-
-internal enum class TestFieldId : FieldId {
-    FIELD_1,
-    FIELD_2,
-    FIELD_3,
-}
